% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_downloader.R
\name{dropbox_downloader}
\alias{dropbox_downloader}
\title{Title}
\usage{
dropbox_downloader(dropbox_link)
}
\arguments{
\item{dropbox_link}{a link to a dropbox *FOLDER* that you *HAVE ACCESS TO*}
}
\value{
a link to the local directory the contents of the dropbox folder have been downloaded to
}
\description{
Title
}
\examples{
dropbox_downloader("https://www.dropbox.com/scl/fo/36jnhue6abszkq2jegjfv/AGRQvghYn6TpCx8uY2KCjZ8?rlkey=4l4xa8nofqtjgojf26xg18p76&dl=1")
^but you need access to the Speckle_Imaging folder in dropbox for this to work
}
